requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
gunicorn==22.0.0
packaging==24.0
sqlparse==0.5.0
typing-extensions==4.11.0
whitenoise==6.6.0
